Avoid listing uploaded files on the reader view that are not yet visible on a page
With this extra condition we more or less ensure that a file that is uploaded
but the page that references it has not yet been approved, is not shown yet
in the list. When the page is also updated on the reader version, the image
will be 'older' than the page so it will be in the list.